requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
aiosqlite
PyYAML
watchdog
openai
agent-framework-core
agent-framework-openai
